+++ src/types/import-data.ts
@@ -1,11 +1,17 @@
 import { Familiars } from "./familiar";
-import { ItemData } from "./inventory-slot";
+import { ItemData } from "./item-data";
 import { Relics } from "./relics";
 
 export interface ImportData {
   presetName: string;
   inventorySlots: ItemData[];
   equipmentSlots: ItemData[];
+
+  /**
+   * Optional fields below.
+   * New fields should be added here for backwards compatibility.
+   */
+
   relics?: Relics;
   familiars?: Familiars;
 }
