+++ src/types/item-data.ts
@@ -1,0 +1,10 @@
+import { EntityData } from "./entity-data";
+
+export interface ItemData extends EntityData {
+  /**
+   * Optional fields.
+   * New fields should be added here for backwards compatibility.
+   */
+  selected?: boolean;
+  slot?: number;
+}
